feat(structure)!: add a placeholder file
Currently the design of the git index (staging area) only permits files to be listed, so committing folders isn't possible. In order to keep the desired folder structure of the repository I added an empty file (placeholder.e) at all folders of the folder structure, which isn't the right way to do it, but it works for laziness purposes.